feat: add test and methods about groups.teams endpoints
