+++ crowdin/crowdin_test.go
@@ -97,7 +97,6 @@ func testClientServices(t *testing.T, c *Client) {
 		"Translations",
 		"TranslationStatus",
 		"MachineTranslationEngines",
-		"Managers",
 		"Screenshots",
 	}
